+++ types/materialize-css/test/autocomplete.test.ts
@@ -1,0 +1,50 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Autocomplete
+const _autocomplete = new M.Autocomplete(elem);
+
+// $ExpectType Autocomplete
+new materialize.Autocomplete(elem);
+// $ExpectType Autocomplete
+const autocomplete = new materialize.Autocomplete(elem, {
+    data: {
+        Apple: null,
+        Google: "https://placehold.it/250x250"
+    },
+    minLength: 3,
+    onAutocomplete(text) {
+        // $ExpectType Autocomplete
+        this;
+        // $ExpectType string
+        text;
+    },
+    sortFunction(a, b, input) {
+        // $ExpectType string
+        a;
+        // $ExpectType string
+        b;
+        // $ExpectType string
+        input;
+        return 0;
+    }
+});
+// $ExpectType void
+autocomplete.updateData({ Microsoft: null });
+// $ExpectType void
+autocomplete.destroy();
+// $ExpectType AutocompleteOptions
+autocomplete.options;
+// $ExpectType Element
+autocomplete.el;
+// $ExpectType boolean
+autocomplete.isOpen;
+
+$(".whatever").autocomplete({
+    data: {
+        Apple: null,
+        Google: "https://placehold.it/250x250"
+    }
+});
+$(".whatever").autocomplete("updateData", { Microsoft: null });
