+++ types/materialize-css/test/carousel.test.ts
@@ -1,0 +1,8 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Carousel
+const carousel = new M.Carousel(elem);
+
+$(".whatever").carousel();
