+++ types/materialize-css/test/character-counter.test.ts
@@ -1,0 +1,16 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType CharacterCounter
+const _characterCounter = new M.CharacterCounter(elem);
+
+// $ExpectType CharacterCounter
+const characterCounter = new materialize.CharacterCounter(elem);
+// $ExpectType void
+characterCounter.destroy();
+// $ExpectType Element
+characterCounter.el;
+
+$(".whatever").characterCounter();
+$(".whatever").characterCounter("destroy");
