+++ types/materialize-css/test/chips.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Chips
+const chips = new M.Chips(elem);
