+++ types/materialize-css/test/collapsible.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Collapsible
+const collapsible = new M.Collapsible(elem);
