+++ types/materialize-css/test/datepicker.test.ts
@@ -1,0 +1,38 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Datepicker
+const _datePicker = new M.Datepicker(elem);
+
+// $ExpectType Datepicker
+new materialize.Datepicker(elem);
+// $ExpectType Datepicker
+const datePicker = new materialize.Datepicker(elem, {
+    defaultDate: new Date(),
+    onSelect(date) {
+        // $ExpectType Datepicker
+        this;
+        // $ExpectType Date
+        date;
+    }
+});
+// $ExpectType void
+datePicker.open();
+// $ExpectType void
+datePicker.setDate(new Date());
+// $ExpectType void
+datePicker.destroy();
+// $ExpectType DatepickerOptions
+datePicker.options;
+// $ExpectType Element
+datePicker.el;
+// $ExpectType boolean
+datePicker.isOpen;
+
+$(".whatever").datepicker();
+$(".whatever").datepicker({ defaultDate: new Date() });
+$(".whatever").datepicker("open");
+$(".whatever").datepicker("destroy");
+$(".whatever").datepicker("setDate", new Date());
+$(".whatever").datepicker("gotoDate", new Date());
