+++ types/materialize-css/test/materialbox.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Materialbox
+const materialbox = new M.Materialbox(elem);
