+++ types/materialize-css/test/parallax.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Parallax
+const parallax = new M.Parallax(elem);
