+++ types/materialize-css/test/pushpin.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Pushpin
+const pushpin = new M.Pushpin(elem);
