+++ types/materialize-css/test/range.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Range
+const range = new M.Range(elem);
