+++ types/materialize-css/test/scrollspy.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType ScrollSpy
+const scrollspy = new M.ScrollSpy(elem);
