+++ types/materialize-css/test/slider.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Slider
+const slider = new M.Slider(elem);
