+++ types/materialize-css/test/taptarget.test.ts
@@ -1,0 +1,6 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType TapTarget
+const taptarget = new M.TapTarget(elem);
