+++ types/materialize-css/test/tooltip.test.ts
@@ -1,0 +1,29 @@
+import * as materialize from "materialize-css";
+
+const elem = document.querySelector('.whatever')!;
+
+// $ExpectType Tooltip
+const _tooltip = new M.Tooltip(elem);
+
+// $ExpectType Tooltip
+new materialize.Tooltip(elem);
+// $ExpectType Tooltip
+const tooltip = new materialize.Tooltip(elem, {
+    inDuration: 300,
+    position: "right"
+});
+// $ExpectType void
+tooltip.open();
+// $ExpectType void
+tooltip.destroy();
+// $ExpectType TooltipOptions
+tooltip.options;
+// $ExpectType Element
+tooltip.el;
+// $ExpectType boolean
+tooltip.isOpen;
+
+$(".whatever").tooltip();
+$(".whatever").tooltip({ html: "<img/>" });
+$(".whatever").tooltip("open");
+$(".whatever").tooltip("destroy");
